cleanup: Fixed up generic tests + made linters pass
